Extract config flow to own module
