+++ homeassistant/components/hue/config_flow.py
@@ -6,7 +6,7 @@ import os
 import async_timeout
 import voluptuous as vol
 
-from homeassistant import config_entries
+from homeassistant import config_entries, flow
 from homeassistant.core import callback
 from homeassistant.helpers import aiohttp_client
 
@@ -41,7 +41,7 @@ def _find_username_from_config(hass, filename):
 
 
 @config_entries.HANDLERS.register(DOMAIN)
-class HueFlowHandler(config_entries.ConfigFlowHandler):
+class HueFlowHandler(flow.ConfigFlowHandler):
     """Handle a Hue config flow."""
 
     VERSION = 1
